Macro conversion WIP
Disclaimer: nothing builds yet, this is just the state of the code as I'm going to bed. You can try building the server module and looking at the output of the println in service.scala to see the generated macro code. Current error is a type mismatch on the marshallers.