implement Wart suppression (#325)
I've created a SupressWarts[A] typeclass with a single method
supressWarts(warts: _*)(a: A) to implement wart supression on several
different language constructs:

- methods (`DefDef`)
- values (`ValDef`)
- classes (`ClassDef`)

Since I'm creating instances of that typeclass for the above mentioned
types, i've left both the typeclass definition and the companion
object inside the def method, but the definition could potentially be
moved outside.

The downside of this implementation is that there's a lot of
boilerplate, but there's no homologue to the `mod""` interpolator for
scalamacros.  I've also thought of implementing the supression on a
structural type `T: {def mods: Modifiers}`, but I find it much more
ugly than current implementation.