Macro conversion (#328)
* Macro conversion WIP

Disclaimer: nothing builds yet, this is just the state of the code as I'm going to bed. You can try building the server module and looking at the output of the println in service.scala to see the generated macro code. Current error is a type mismatch on the marshallers.

* Various bug fixes

There remains a bug related to constructor generation, and we need to bring back the warnings and wart-remover suppressions

* Fixed missing-constructor bug

All tests now pass except the idlGen module which remains to be converted.

We still need some cleanups to the `@service` handling code including restoring the warnings suppression and moving the context-dependent inner functions elsewhere if possible.

* Misc cleanup of service macro, docs and deprecation fixes.

Renamed a few things in the `@service` macro handler for clarity, added todos for wart-suppression and others, removed deprecated annotation in tests and docs, updated scalameta reference and removed free-config reference in docs, fixed Scala version in Travis, bumped minor project version.

* implement Wart suppression (#325)

I've created a SupressWarts[A] typeclass with a single method
supressWarts(warts: _*)(a: A) to implement wart supression on several
different language constructs:

- methods (`DefDef`)
- values (`ValDef`)
- classes (`ClassDef`)

Since I'm creating instances of that typeclass for the above mentioned
types, i've left both the typeclass definition and the companion
object inside the def method, but the definition could potentially be
moved outside.

The downside of this implementation is that there's a lot of
boilerplate, but there's no homologue to the `mod""` interpolator for
scalamacros.  I've also thought of implementing the supression on a
structural type `T: {def mods: Modifiers}`, but I find it much more
ugly than current implementation.

* migrate idlgen to scalamacros (#326)

* create a MacrosUtil class

Adapted from Scalameta Utils

* modify scalaparser to work with scalamacros

* add scala compiler as a dependency

* add monocle dependency for `internal` project

* create monocle optics to work with Scala AST

Currently i've only created the ones I inmediately needed.

* delete MacrosUtil

functionality is mostly substituted by AstOptics

* create a prism for focusing on case classes

* add params optic & hasAnnotation combinator

* move AST prisms to the internal `ast` object

also, create optics for extracting name of request & response types correctly

* migrate Model to scalamacros

- delete scalameta references
- change scalameta AST references to scala internals

* make ScalaParser work with scalamacros

* correctly match the IDL type

* create optic to extract the IDL type from an @rpc annotated method

* simplify getRequestsFromService

* needed tweak to SBT

* fix StreamingType parsing

* migrate IdlGenerator to scalamacros

* create a single instance of Toolbox

And make dependents depend on it instead of passing the instance
around (which breaks path dependent typing)

* delete unneeded imports

* migrate AvroIdlGenerator

* migrate ProtoIDLGenerator

* use Trees instead of Strings

* make IdlGenTests pass

* address comments in PR

* fix travis error on 2.11. first try

* fix travis error on 2.11. second try

* avoid the use of `.head`

* fix the bug with scala 2.11 in travis

* fix idlgen-sbt plugin tests

* :fire: head

* fix findAnnotation

* merge for & collect

* create extractors for matching on complex types