Replace macros with codegen - POC/WIP
Created a compiler plugin (not an sbt plugin) to generate managed sources from the @service definitions
(here @serviceCodegen to preserve the originals) without using macros but by reusing most of the macro's
implementations. Note that this requires a 2-subproject setup so that the app's compilation can see the
sources generated by the protocol's compilation. This is shown in example/codegen.