Upgrade to latest snapshots of pbdirect and skeuomorph
This version of pbdirect requires all fields to be annotated with a
`@pbIndex` and only allows messages to be encoded, not individual
fields.

This version of skeuomorph generates code that includes those
annotations.

It's a huge diff but it's mostly just me going around and adding the
annotation everywhere.

Note: the CI build will fail because it's relying on locally built
snapshots.