Tweak test data
Reordered the coproduct choices in `case class RequestCoproductNoInt` so
we can provide a valid Avro default value (the default value must
correspond to the first type in the union).

Reordered all the other coproducts for consistency.

The tests still fail, but they now fail in a way that will be fixed by
https://github.com/sksamuel/avro4s/pull/439