Automatically register muSrcGen as a sourceGenerator
It doesn't really make sense *not* to add the task to the list of
sourceGenerators, as the list of generated source files would not be
added to `managedSources`. That means even if you ran `muSrcGen`
manually and then ran `compile`, the generated source files would not be
compiled.

Updated the plugin to automatically register the `muSrcGen` task as a
sourceGenerator, and updated the misleading guidance in the
documentation.