Automatically register muSrcGen as a sourceGenerator (#813)
* Automatically register muSrcGen as a sourceGenerator

It doesn't really make sense *not* to add the task to the list of
sourceGenerators, as the list of generated source files would not be
added to `managedSources`. That means even if you ran `muSrcGen`
manually and then ran `compile`, the generated source files would not be
compiled.

Updated the plugin to automatically register the `muSrcGen` task as a
sourceGenerator, and updated the misleading guidance in the
documentation.

* Fix/update all the sbt plugin scripted tests

A lot of them were not really testing anything. They were running
`compile`, but that task wasn't compiling any files so it was trivially
succeeding.

Others just needed some tidy-up because they hadn't been touched for a
long time.