Add support for tracing of unary calls on server side
The `@service` macro now generates a new method:

```
def bindServiceWithTracing[F[_]](entrypoint: natchez.EntryPoint[F])
  (implicit alg: MyService[Kleisli[F, natchez.Span[F], *]], ...)
```

This is similar to the existing `bindService` method, but when handling
a call it creates a Span and uses it to run a Kleisli produced by the
algebra.

If the request contains tracing headers, the call handler will create a
child span to continue the trace. Otherwise it will create a root span.
The name of the span is the fully qualified name of the RPC method, e.g.
"com.example.Greeter/SayHello".

The user writes their service using `F: Trace: Async`, so they can
create child spans as necessary. Within the service code there is no
need to know that F is actually a Kleisli. See the Natchez docs for more
details.