Distributed tracing with Natchez (#866)
* Add natchez-core as a core dependency

* Add support for tracing of unary calls on server side

The `@service` macro now generates a new method:

```
def bindServiceWithTracing[F[_]](entrypoint: natchez.EntryPoint[F])
  (implicit alg: MyService[Kleisli[F, natchez.Span[F], *]], ...)
```

This is similar to the existing `bindService` method, but when handling
a call it creates a Span and uses it to run a Kleisli produced by the
algebra.

If the request contains tracing headers, the call handler will create a
child span to continue the trace. Otherwise it will create a root span.
The name of the span is the fully qualified name of the RPC method, e.g.
"com.example.Greeter/SayHello".

The user writes their service using `F: Trace: Async`, so they can
create child spans as necessary. Within the service code there is no
need to know that F is actually a Kleisli. See the Natchez docs for more
details.

* Add scala-collection-compat

* Format sbt files

* Support for adding extra request headers to unary call

* Support sending custom headers from client in fs2 streaming calls

* Give fs2 calls access to the request headers on server side

We need this to support distributed tracing.

* Helper methods for Kernel <-> Metadata conversion

* Add macro-generated MyService.tracingClient function

```
def tracingClient[F: ConcurrentEffect: ContextShift](
  channelFor: ChannelFor,
  channelConfigList: List[ManagedChannelConfig],
  options: CallOptions
): Resource[F, MyService[Kleisli[F, Span[F], *]]]
```

Only support unary calls for now.

* Tracing of fs2 client-streaming calls

Implementations for both client and server side.

* Get the tests compiling again

* Refactor to use helper method

* Implement tracing of FS2 server-streaming endpoints

* Implement tracing of FS2 bidirectional-streaming endpoints

* Implement client-side tracing of Monix streaming endpoints

* Implement server-side tracing of Monix streaming endpoints

* Rename and move some things for consistency

* Tweak package layout and object naming

* Tighten up a couple of context bounds

* Implement tracingClientFromChannel, unsafeTracingClient{fromChannel}

* Remove dead val

* Tests for tracing

Monix client-streaming with tracing compiles but doesn't work. Rather
worrying... feels like Observable and Kleisli don't play nicely.

* Use a real server and client in RPC tracing tests

Make the server and client communicate over a real channel, not an
in-process channel. For some reason this fixes the mysterious Monix
empty Observable issue.

* Write how-to guide for distributed tracing