Make the OkHttp and Netty tests actually use OkHttp/Netty
grpc-java chooses the client implementation by finding all `Provider`s
on the classpath and picking the one with the highest priority. The
`tests` module has both `grpc-netty` and `grpc-okhttp` on the classpath,
so it will always choose the Netty implementation unless we specify
otherwise.

Added a couple of parameters to `ManagedChannelInterpreter`'s
constructor to allow us to inject the appropriate implementation in
tests, and updated the tests accordingly.