Add a TaggedDecimal type to replace Shapeless tagging
`TaggedDecimal[P <: Int, S <: Int]` is a BigDecimal tagged with the
precision and scale extracted from its corresponding `decimal` logical
type in the Avro schema.

Implemented as an opaque type, with a smart constructor that performs a
couple of sanity checks on the precision and scale of the supplied
BigDecimal.