Remove dependency from tests module on config module
We can't cross-build the config module because it depends on pureconfig,
so we need to remove this inter-module dependency in order to
cross-build the tests module.

Moved the relevant tests into the config module.

Also fixed one of the tests that didn't make much sense, and added a
missing config file.