Switch from cross-building to sbt-crossproject
We have most modules targeting Scala 2.13 + Scala 3, a few modules
(config, benchmark-vnext, haskell-integration-tests) that only target
2.13, and now I want to upgrade the microsite to target Scala 3.

Unfortunately that seems to be impossible using sbt's built-in
scalaVersion/crossScalaVersions mechanism. Whatever I do, I get errors
about both _2.13 and _3 versions of dependencies being on the classpath.

Switched to sbt-crossproject, which makes it really easy to handle
situations like this.

It also allows more parallelisation, as the 2.13 and 3 versions of a
module will build in parallel.